backend/requirements.txt
asgiref==3.5.2
device-detector==5.0.1
Django==3.2.13
django-cors-headers==3.12.0
django-jazzmin==2.5.0
django-jwt-extended==0.1.1
django-user-agents==0.4.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
PyJWT==1.7.1
pytz==2022.1
PyYAML==6.0
regex==2022.4.24
sqlparse==0.4.2
tzdata==2022.1
ua-parser==0.10.0
user-agents==2.2.0
Werkzeug==2.1.2

# FOR VAD TASKS
huggingface_hub>=0.0.6
diart==0.2.1
hyperpyyaml>=0.0.1
joblib>=0.14.1
numpy>=1.17.0
noisereduce==2.0.0
librosa==0.8.1
packaging
pandas==1.4.0
pre-commit>=2.3.0
PyAudio==0.2.11
scikit-learn==1.0.1
scipy>=1.4.1
sentencepiece>=0.1.91
SoundFile==0.10.3.post1
torch>=1.8.0,<=1.11.0
pydub>=0.25.1
torchaudio>=0.8.0,<=0.11.0
tqdm>=4.42.0
-e git+https://github.com/speechbrain/speechbrain/@59999b1db4a1057fcd7b1f496cecda6c58e37289#egg=speechbrain
https://github.com/pyannote/pyannote-audio/archive/develop.zip
